+++ packages/eyes-storybook/src/generateConfig.js
@@ -50,9 +50,6 @@ function generateConfig({argv = {}, defaultConfig = {}, externalConfigParams = [
   if (result.storyDataGap === undefined) {
     result.storyDataGap = result.testConcurrency;
   }
-  if (result.saveNewTests === undefined) {
-    result.saveNewTests = true;
-  }
   return result;
 }