+++ packages/utils/src/general.ts
@@ -74,13 +74,12 @@ export function toString(object: Record<PropertyKey, any>): string {
   return `${this.constructor.name} ${JSON.stringify(object, null, 2)}`
 }
 
-export function toMany(object: [] | number, config?: [manyCase: string, singleCase: string]): string {
-  const count = types.isArray(object) ? object.length : types.isNumber(object) ? object : 0;
-  const isMany = count > 1;
-  let res = isMany ? 's' : '';
-  if (config){
-    res = isMany ? config[0] || '' : config[1] || '';
+export function pluralize(object: [] | number, config?: [manyCase: string, singleCase: string]): string {
+  const count = types.isArray(object) ? object.length : object
+  const isMany = count > 1
+  let res = isMany ? 's' : ''
+  if (config) {
+    res = isMany ? config[0] : config[1]
   }
-  return res;
+  return res
 }
-